refactor: change name of main file
